+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 Unreleased
 ----------
 
+Changed:
+- CRN.equilibrate now accepts 2D initial conditions with specialized CRN.params
+
 Fixed:
 - CRN.parametrize_for is now more careful when substituting specialized parameters in expressions